+++ Interface/AddOns/nibRealUI/Infobar/Bar.lua
@@ -8,7 +8,6 @@ local LDB = _G.LibStub("LibDataBroker-1.1")
 local qTip = _G.LibStub("LibQTip-1.0")
 local Aurora = _G.Aurora
 local Base = Aurora.Base
-local frameColor = Aurora.Color.frame
 
 -- RealUI --
 local RealUI = private.RealUI
@@ -707,7 +706,7 @@ function Infobar:CreateBar()
     end)
 
     -- Stripes
-    Base.SetBackdrop(frame, frameColor, db.bgAlpha)
+    Base.SetBackdrop(frame, nil, db.bgAlpha)
     local tex = frame:CreateTexture(nil, "BACKGROUND", nil, 1)
     tex:SetTexture([[Interface\AddOns\nibRealUI\Media\StripesThin]], true, true)
     tex:SetAlpha(db.bgAlpha * 0.6)
@@ -809,7 +808,7 @@ function Infobar:SettingsUpdate(setting, block)
         end
         block:OnEvent("SettingsUpdate")
     elseif setting == "bgAlpha" then
-        Base.SetBackdropColor(self.frame, frameColor.r, frameColor.g, frameColor.b, db.bgAlpha)
+        Base.SetBackdropColor(self.frame, nil, db.bgAlpha)
         self.frame.tex:SetAlpha(db.bgAlpha * 0.6)
         self.frame.watch:UpdateColors()